feat(Home,SiginIn,Header) 홈 background image 추가, 로고 및 회원가입 클릭 시 이동 추가
