+++ Front-end/schedular/src/AppRouter.tsx
@@ -1,4 +1,4 @@
-import {BrowserRouter, Route, Routes, Navigate} from 'react-router-dom';
+import {Route, Routes} from 'react-router-dom';
 import Home from './Pages/Home';
 import SignUp from './Pages/SignUp';
 import DemoApp from './Pages/FullCalendar';
@@ -6,13 +6,11 @@ import DemoApp from './Pages/FullCalendar';
 const AppRouter = () =>{
     return (
         <>
-            <BrowserRouter>
-                <Routes>
-                    <Route path="/" element={<Home />}/>
-                    <Route path="/signup" element={<SignUp />}/>
-                    <Route path="/calendar" element={<DemoApp />}/>
-                </Routes>
-            </BrowserRouter>
+            <Routes>
+                <Route path="/" element={<Home />}/>
+                <Route path="/signup" element={<SignUp />}/>
+                <Route path="/calendar" element={<DemoApp />}/>
+            </Routes>
         </>
     )
 }
