+++ Front-end/schedular/src/Assets/Images/eclass_logo.png
