+++ Front-end/schedular/src/Assets/Images/logo.png
