+++ Front-end/schedular/src/Components/Header.tsx
@@ -1,39 +1,24 @@
-import SignIn from "../SignIn";
-import styled from "styled-components";
-import eclass_logo from "../Images/eclass_logo.png"
+import SignIn from './SignIn';
+import { Link } from 'react-router-dom';
+import styled from 'styled-components';
+import eclass_logo from '../Assets/Images/eclass_logo.png'
 
-   
-const Header = () => {
-  return(
-    <HeaderWapper>
-        <EclassLogo src={eclass_logo} alt="eclass_logo"/>
-        <SignIn></SignIn>
-        <a href="/SignUp">
-            <SignUpButton>회원가입</SignUpButton>
-        </a>
-      </HeaderWapper>
-  )
-}
-      
 const HeaderWapper = styled.div`
   height: 100px;
   display: flex;
 `;
-
-const EclassLogo = styled.img`
+const EclassLogo = styled(Link)`
   position: relative;
   top: 0;
   botton: 0;
   margin: auto 10px auto 20px;
   height: 42px;
-  width: 240px;  
+  width: 240px;
+  background-image: url(${eclass_logo});
+  background-size: cover;
 `;
-
-
-const SignUpButton = styled.button`
+const SignUpButton = styled(Link)`
   margin 33px 0;
-  height: 35px;
-  width: 80px;
   text-align: center;
   color: #fff;
   padding: 7px 16px;
@@ -43,11 +28,15 @@ const SignUpButton = styled.button`
   font-size: 12px;
 `;
 
-const BgImage = styled.div`
-  height: 100%;
-  background-color: orange;
-  background-image: url(../Images/IMG_4174.jpeg);
-`
+const Header = () => {
 
+  return(
+    <HeaderWapper>
+      <EclassLogo to='/'></EclassLogo>
+      <SignIn></SignIn>
+      <SignUpButton to='/signup'>회원가입</SignUpButton>
+    </HeaderWapper>
+  )
+}
 
 export default Header;
