+++ Front-end/schedular/src/Pages/Home.tsx
@@ -1,37 +1,26 @@
-import Header from "../Components/Header";
-import styled from "styled-components";
+import styled from 'styled-components';
+import background_image from '../Assets/Images/IMG_4174.jpeg'
 
-
-
+const BgImage = styled.div`
+  min-height: 800px;
+  background-size: cover;
+`;
+const Content = styled.div`
+`
 
 const Home = () => {
 
   return (
     <>
-      <Header></Header>
       <Content>
-        <BgImage>
-          background image
+        <BgImage style={{ backgroundImage: `url(${background_image})`}}>
           <div>
-            내 할 일 보기
-            내 강의실
+            
           </div>
         </BgImage>
       </Content>
     </>
   )
 }
-
-
-
-const BgImage = styled.div`
-  height: 100%;
-  background-color: orange;
-  background-image: url(../Images/IMG_4174.jpeg);
-`;
-
-const Content = styled.div`
-`
-
 
 export default Home
